requirements.txt
attrs==21.2.0
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
Deprecated==1.2.12
Flask==1.1.2
Flask-Compress==1.9.0
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyGithub==1.55
Pygments==2.9.0
PyJWT==2.1.0
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.2.4
requests==2.25.1
rich==10.1.0
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
